+++ src/main/java/com/pekka/moni/transaction/dto/TransactionDateSpanResponse.java
@@ -2,6 +2,7 @@ package com.pekka.moni.transaction.dto;
 
 import com.pekka.moni.transaction.Transaction;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-public record TransactionDateSpanResponse(List<Transaction> transactions, Double sum) {}
+public record TransactionDateSpanResponse(List<Transaction> transactions, BigDecimal sum) {}
